+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added [Tree component](src/resources/components/interact/Tree.tsx)
-* Change [Aside partial](src/resources/components/partials) to show the tree component
+* Change [Aside partial](src/resources/components/partials/Aside.tsx) to show the tree component
+* Aside is now [OffCanvas](src/resources/components/interact/OffCanvas.tsx) in Holy
 
 ## [0.1.1] - 2026-06-18